regedit: Introduce a new API to build the dialogs.
This helps make new dialogs easier to create, because it provides some
common building blocks and a consistent way to deal with user input.

Dialogs are divided into sections that stack vertically, and common
sections for typical UI things like text boxes, option lists, and
buttons are provided.

The old dialogs are rewritten to use this API.

Signed-off-by: Chris Davis <cd.rattan@gmail.com>
Reviewed-by: Andreas Schneider <asn@samba.org>
Reviewed-by: Michael Adam <obnox@samba.org>